extract buildel types
